+++ controllers/transactions/response/json.go
@@ -22,7 +22,7 @@ type Transaction struct {
 	Price          float64              `json:"price"`
 	Product        string               `json:"product"`
 	Payment_method string               `json:"payment_method"`
-	Point          float32              `json:"point"`
+	Point          float64              `json:"point"`
 	Status         string               `json:"status"`
 	Description    string               `json:"description"`
 	UserID         uint                 `json:"user_id"`
@@ -40,7 +40,7 @@ type TransactionInAdmin struct {
 	Price          float64              `json:"price"`
 	Product        string               `json:"product"`
 	Payment_method string               `json:"payment_method"`
-	Point          float32              `json:"point"`
+	Point          float64              `json:"point"`
 	Status         string               `json:"status"`
 	Description    string               `json:"description"`
 	UserID         uint                 `json:"user_id"`
