+++ controllers/vouchers/response/json.go
@@ -8,7 +8,7 @@ type Voucher struct {
 	ID      uint    `json:"id" gorm:"primaryKey"`
 	Product string  `json:"product"`
 	Benefit string  `json:"benefit"`
-	Cost    float32 `json:"cost"`
+	Cost    float64 `json:"cost"`
 }
 
 func FromDomain(domain vouchers.Domain) Voucher {
